+++ src/main/java/fi/csc/microarray/client/visualisation/methods/gbrowser/track/QualityCoverageTrack.java
@@ -154,7 +154,7 @@ public class QualityCoverageTrack extends Track {
 
 			for (int i = 0; i < quality.length(); i++) {
 
-				int refIndex = cigar.getReferenceIndex(i);
+				long refIndex = cigar.getReferenceIndex(i);
 
 				if (refIndex == -1) {
 					//Skip insertions
